backend/requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
attrs==20.3.0
Django==3.1.7
factory-boy==3.2.0
Faker==6.5.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
promise==2.3
psycopg2-binary==2.8.6
py==1.10.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
